+++ memberaudit_securegroups/__init__.py
@@ -2,8 +2,4 @@
 Init module
 """
 
-# Django
-from django.utils.translation import gettext_lazy as _
-
 __version__ = "0.6.1"
-__title__ = _("Member Audit Secure Groups")